+++ src/api/path/index.ts
@@ -1,4 +1,5 @@
 const PRODUCT = {
+  OPTION: (productId: string) => `/v1/products/${productId}/options`,
   DETAIL: (productId: string) => `/v1/products/${productId}/detail`,
 } as const;
